Update Renderer and resources
The renderer engine has been updated to handle all responsability rather
than delegating it. Now it has all control and can be expanded in order
to work for multiple shaders in diferent sprites.

The resources no longer needs the loader attribute. The new FileHandle
class was created to work with the Resources and the FileHandlLoader can
load any file in the resource manager.