Add Buffers. Add onUpdate method for the scene.
Buffers will help with passing non-static values and scene dependent
values to prefabs. As an example, receive an EntityBuffer as attribute,
then assign that attribute in the resources and add some entity to that
buffer. Finally, in the onCreate method, that buffer will be assigned
with an Entity inside.

The onUpdate method is intended to be used to update resources and
general systems and game managers. In general it's used to do things
that do not feel right to do in components.