Add coreUtil overrideFileWith and export manifest
The CoreUtil now provides the overrideFileWith method with replaces the
content of a file with new content and creates it if it doesn't exists.
(It uses the FileHandle).

In development, now you can create the package manifest java files just
by setting a boolean when using a package with the automatic loader.